alx_travel_app/requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
certifi==2019.11.28
chardet==3.0.4
dbus-python==1.2.16
distro==1.4.0
distro-info===0.23ubuntu1
django==4.2.21
django-cors-headers==4.4.0
django-environ==0.11.2
djangorestframework==3.15.2
drf-yasg==1.21.10
greenlet==1.1.2
idna==2.8
inflection==0.5.1
mysql-connector-python==9.0.0
mysqlclient==2.1.0
numpy==1.15.0
packaging==25.0
pep8==1.7.1
psutil==5.9.0
pycodestyle==2.8.0
Pygments==2.3.1
PyGObject==3.36.0
python-apt==2.0.0+ubuntu0.20.4.7
python-dotenv==1.0.1
pytz==2025.2
PyYAML==5.3.1
requests==2.22.0
requests-unixsocket==0.2.0
six==1.14.0
SQLAlchemy==1.4.32
sqlparse==0.5.3
ssh-import-id==5.10
typing-extensions==4.13.2
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==1.25.8
